+++ sdk/dashboard/azure-mgmt-dashboard/azure/mgmt/dashboard/_configuration.py
@@ -19,7 +19,7 @@ if TYPE_CHECKING:
     from azure.core.credentials import TokenCredential
 
 
-class DashboardManagementClientConfiguration(Configuration):
+class DashboardManagementClientConfiguration(Configuration):  # pylint: disable=too-many-instance-attributes
     """Configuration for DashboardManagementClient.
 
     Note that all parameters used to create this instance are saved as instance
@@ -27,8 +27,11 @@ class DashboardManagementClientConfiguration(Configuration):
 
     :param credential: Credential needed for the client to connect to Azure.
     :type credential: ~azure.core.credentials.TokenCredential
-    :param subscription_id: Gets subscription credentials which uniquely identify Microsoft Azure subscription. The subscription ID forms part of the URI for every service call.
+    :param subscription_id: The ID of the target subscription.
     :type subscription_id: str
+    :keyword api_version: Api Version. Default value is "2022-08-01". Note that overriding this
+     default value may result in unsupported behavior.
+    :paramtype api_version: str
     """
 
     def __init__(
@@ -38,6 +41,8 @@ class DashboardManagementClientConfiguration(Configuration):
         **kwargs: Any
     ) -> None:
         super(DashboardManagementClientConfiguration, self).__init__(**kwargs)
+        api_version = kwargs.pop('api_version', "2022-08-01")  # type: str
+
         if credential is None:
             raise ValueError("Parameter 'credential' must not be None.")
         if subscription_id is None:
@@ -45,7 +50,7 @@ class DashboardManagementClientConfiguration(Configuration):
 
         self.credential = credential
         self.subscription_id = subscription_id
-        self.api_version = "2021-09-01-preview"
+        self.api_version = api_version
         self.credential_scopes = kwargs.pop('credential_scopes', ['https://management.azure.com/.default'])
         kwargs.setdefault('sdk_moniker', 'mgmt-dashboard/{}'.format(VERSION))
         self._configure(**kwargs)