+++ sdk/dashboard/azure-mgmt-dashboard/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Release History
 
+## 1.0.0 (2022-07-19)
+
+**Features**
+
+  - Added operation group PrivateEndpointConnectionsOperations
+  - Added operation group PrivateLinkResourcesOperations
+  - Model ManagedGrafanaProperties has a new parameter api_key
+  - Model ManagedGrafanaProperties has a new parameter deterministic_outbound_ip
+  - Model ManagedGrafanaProperties has a new parameter outbound_i_ps
+  - Model ManagedGrafanaProperties has a new parameter private_endpoint_connections
+  - Model ManagedGrafanaProperties has a new parameter public_network_access
+  - Model ManagedGrafanaUpdateParameters has a new parameter properties
+
 ## 1.0.0b2 (2022-04-13)
 
 **Features**